feat : img2txt -> gpt 분리 (gpt_api.py)에 저장 .env 파일에 prompt 분리
