+++ app/services/img2txt.py
@@ -1,8 +1,7 @@
-import os, gc, torch, requests, openai
+import os, gc, torch, requests
 import requests
 from io import BytesIO
 from PIL import Image
-from transformers import BlipProcessor, BlipForConditionalGeneration
 from core.config import settings
 import logging
 
@@ -12,62 +11,6 @@ class ImageToText:
     def __init__(self, blip_model, processor):
         self.blip_model = blip_model
         self.processor = processor
-        self.client = openai.OpenAI(api_key=settings.OPENAI_API_KEY)
-    
-    def parse_gpt_output(self, text: str) -> tuple[str, list[str]]:
-        prompt = ""
-        items = []
-
-        try:
-            lines = text.strip().split("\n")
-            reading_prompt = False
-            reading_items = False
-
-            for line in lines:
-                if line.lower().startswith("prompt:"):
-                    prompt = line.split(":", 1)[1].strip()
-                    reading_prompt = True
-                    reading_items = False
-                elif "recommended items" in line.lower():
-                    reading_items = True
-                    reading_prompt = False
-                elif reading_prompt and prompt == "":
-                    prompt = line.strip()
-                elif reading_items and line.strip().startswith(("-", "•")):
-                    items.append(line.strip("-• ").strip())
-
-            if not prompt:
-                raise ValueError("Prompt not found.")
-            if len(items) < 3:
-                raise ValueError("Too few items extracted.")
-
-            return prompt, items
-        except Exception as e:
-            logger.error(f"GPT 응답 파싱 실패: {e}")
-            logger.info(f"원본 GPT 응답:\n{text}")
-            
-            fallback_prompt = text.split("Prompt:")[-1].split("Recommended Items:")[0].strip() if "Prompt:" in text else ""
-            logger.info("기본 추천 키워드로 대체합니다.")
-            return fallback_prompt, [
-                "mouse", 
-                "desk mat", 
-                "mechanical keyboard", 
-                "LED desk lamp", 
-                "potted plant"
-            ]
-
-    # Prompt 정리(불필요한 단어 제거)
-    def clean_prompt(self, prompt: str) -> str:
-        prompt = prompt.replace("wired", "wireless")
-        prompt = prompt.replace("cables", "no visible cables")
-        prompt = prompt.replace("clutter", "clean and organized")
-        prompt = prompt.replace("noisy background", "neutral background")
-        # 쓰레기 제거
-        trash_keywords = ["trash", "mess", "noodles", "instant ramen", "wires", "tangled"]
-        for word in trash_keywords:
-            if word in prompt:
-                prompt = prompt.replace(word, "")
-        return prompt.strip()
 
     def generate_text(self, url: str):
         if not url:
@@ -84,67 +27,13 @@ class ImageToText:
 
             logger.info(f"Caption: {caption}")
 
-            # GPT-4o Prompts 생성
-            response = self.client.chat.completions.create(
-            model = "gpt-4o",
-            messages = [
-                {
-                    "role": "system",
-                    "content": (
-                        "You are an expert prompt engineer and interior stylist for text-to-image models like Stable Diffusion XL.\n"
-                        "Your job is to take a simple image caption describing a basic desk layout, and do the following:\n\n"
-                        "1. Rewrite the caption into a vivid, photorealistic prompt that imagines a beautifully styled and enhanced version of the same desk setup.\n"
-                        "   ✨ You MUST preserve the original objects described in the caption (e.g., a laptop and monitor), and build on top of them with lighting, layout, and atmosphere enhancements.\n"
-                        "   The rewritten prompt MUST be under 70 tokens.\n\n"
-                        "2. Then, separately, suggest exactly 5 practical and stylish desk accessories that match the upgraded prompt.\n"
-                        "   Each item must be a short, search-friendly **keyword**, not a full sentence or description.\n"
-                        "   Use lowercase, concise nouns like: 'desk lamp', 'monitor stand', 'plant'.\n\n"
-                        "**Output Format (strict):**\n"
-                        "Prompt:\n<your styled prompt here (under 70 tokens)>\n\n"
-                        "Recommended Items:\n"
-                        "- item 1\n"
-                        "- item 2\n"
-                        "- item 3\n"
-                        "- item 4\n"
-                        "- item 5\n\n"
-                        "**Rules:**\n"
-                        "1. Do NOT remove or replace original caption objects (e.g., don't remove the laptop/monitor if they were mentioned).\n"
-                        "2. Do NOT change the scene type (e.g., do NOT move it to a coffee shop or bedroom).\n"
-                        "3. Only improve the style and detail of the existing scene.\n"
-                        "4. The generated scene should be captured from a clean, centered, front-facing camera angle — as if looking directly at the desk from eye level. Avoid tilted or top-down views.\n"
-                        "5. You MUST generate 5 recommended items. Never generate fewer than 5.\n"
-                        "6. If necessary, you may include objects from the original caption (e.g., laptop or monitor) as part of the 5 items.\n"
-                        "7. The recommended items must be **search-optimized keywords** — avoid long phrases or descriptive sentences."
-                    )
-                },
-                {
-                    "role": "user",
-                    "content": (
-                        f"Here is a caption of a desk scene:\n\n"
-                        f"Caption: {caption}\n\n"
-                        "Please follow all instructions strictly.\n"
-                        "Rewrite the caption into a vivid, styled image generation prompt.\n"
-                        "Then suggest exactly 5 short, search-optimized desk accessory keywords as described above.\n"
-                        "Return the result in the exact output format requested."
-                    )
-                }
-            ],
-            temperature=0.6,
-            max_tokens=300  # 🔼 추천: 70은 너무 작음 (prompt + list까지 포함 못함)
-            )
             
             gc.collect()
             torch.cuda.empty_cache()
-            logger.info("Prompt 생성 완료")
-            # generated_prompt = response.choices[0].message.content
-            # generated_prompt = self.clean_prompt(generated_prompt)
-            generated_prompt = response.choices[0].message.content
-            cleaned_prompt, items = self.parse_gpt_output(generated_prompt)
-            cleaned_prompt = self.clean_prompt(cleaned_prompt)
+            logger.info("Caption 생성 완료")
             
             # 여기서 추천 아이템은 어떤 식으로 뽑아올 지 좀 더 고민
-            return cleaned_prompt, items
-        
+            return caption
         except Exception as e:
             logger.error(f"generate_text() 실패: {e}")
             return None
